move api related functions to api.js
